rayton_tg_service/requirements.txt
telethon==1.36.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
